+++ docs/CHANGELOG.md
@@ -11,7 +11,7 @@ published.
 
 ## [0.2.0] - 2025-07-26
 ### Added
-- New `website_v4` theme under `clients/baattilsyn/website`.
+ - New `website_v4` theme under `themes/baattilsyn/website`.
 - Updated test themes structure.
 
 ## [0.1.0] - 2025-07-25